requirements.txt
flask==3.0.0
gunicorn==21.2.0
aiohttp==3.9.1
pyairtable==2.1.0
schedule==1.2.0
requests==2.31.0
twilio==8.10.0
python-dotenv==1.0.0
tenacity==8.2.3
pytest==7.4.3
pytest-asyncio==0.21.1
